# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Kingston Care Connect API
  description: |
    Public API for accessing community services in Kingston, Ontario.

    **Base URL:** `https://kingston-care-connect.vercel.app/api/v1`

    ## Authentication
    - Public endpoints (services) require no authentication.
    - Analytics endpoints require a valid session cookie (Magic Link login).
  version: 1.0.0
  contact:
    email: support@kingstoncare.ca
servers:
  - url: https://kingston-care-connect.vercel.app/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local Development
paths:
...
  /analytics:
    get:
      summary: Get Analytics (Authenticated)
      description: |
        Retrieve analytics data for services. Requires authentication via Magic Link.
        Partners can view aggregated view/click counts for their services.
      operationId: getAnalytics
      tags:
        - Analytics
      security:
        - cookieAuth: []
      parameters:
        - name: service_id
          in: query
          description: Filter by specific service ID
          schema:
            type: string
        - name: days
          in: query
          description: Number of days to look back (default 30, max 90)
          schema:
            type: integer
            default: 30
            maximum: 90
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AnalyticsEntry"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session cookie (set after Magic Link login)
  schemas:
...
    AnalyticsEntry:
      type: object
      properties:
        service_id:
          type: string
        views:
          type: integer
          description: Number of detail page views
        clicks:
          type: integer
          description: Number of website/call clicks
        period_days:
          type: integer
          description: Lookback period in days
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"